Kind of had to do all the rest of the moving at once to keep the project running (the rest of the owl)
