+++ gherkin/model/__init__.py
@@ -1,0 +1,4 @@
+from gherkin.model.robot import Robot, RobotLimits
+from gherkin.model.world import World
+
+__all__ = ["Robot", "RobotLimits", "World"]
